Add recent order service
This code was copied over from a programming challenge I completed. It is obfuscated from what the original challenge was and the commits replace with this one in order to preserve privacy for the company. This is a solid starter project Spring Boot example so I want to keep it around.